requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-dotenv==1.0.0
chromadb==0.4.15
numpy==1.26.4
faiss-cpu==1.8.0
setuptools
wheel
sentence-transformers
huggingface-hub